initial

+++ findMedianFromAStreamOFData.cpp
@@ -16,7 +16,7 @@ double findMedian(vector <int> v){
         return (v[n / 2] + v[n / 2 - 1]) / 2.0;
     }
     else{
-        return v[n / 2];
+        return  (double) v[n / 2];
     }
 }
